Added size to folders and refactored the integration tests.
