+++ src/main/java/com/grongo/cloud_storage_app/models/items/dto/FileDto.java
@@ -11,6 +11,5 @@ import lombok.experimental.SuperBuilder;
 @AllArgsConstructor
 @SuperBuilder
 public class FileDto extends ItemDto{
-    private Long size;
     private String fileType;
 }
