+++ src/test/java/com/grongo/cloud_storage_app/integrationTests/AuthenticationIT.java
@@ -18,13 +18,12 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
 import static org.assertj.core.api.Assertions.*;
 import java.util.Optional;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static com.grongo.cloud_storage_app.testUtils.TestUtils.*;
+import static com.grongo.cloud_storage_app.TestUtils.*;
 
 @AutoConfigureMockMvc
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
